characters: use single tiles to build animation
Since the sprite can have images of different sizes as part of a
unique animation, it's not possible to identify a single size and
start from the first frame moving to a direction of the given size
multiplied by the frame number. We must build the animation as a
sequence of single images, regardless of their sizes or position
in the sprite.
With this commit this works only for frames of the same size